Move attatchment contexts under vocabulary context
